requirements.txt
MLB-StatsAPI==1.9.0
requests==2.32.5
